% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Correlate the z-scored values to the historic oni-intensity.}
\usage{
correlate(gage_list, project_directory)
}
\arguments{
\item{gage_list}{The gages used in vector form.}

\item{project_directory}{The directory used to save output files.}
}
\value{
A dataframe of correlated values using the sum, min, max, and mean z-scored values.
}
\description{
Correlate the z-scored values to the historic oni-intensity.
}
