% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Detrend}
\usage{
detrend(x, y)
}
\arguments{
\item{x}{The date vector or column for the discharge data.}

\item{y}{The discharge values to be detrended.}
}
\value{
A list of the detrended data and the linear regression model.
}
\description{
Perform a linear regression on the discharge to remove trends.
}
