% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_parameters.R
\name{discharge_parameters}
\alias{discharge_parameters}
\title{Calculate the annual discharge, the average daily discharge, the 7 day minimum discharge, and the 15 day maximum discharge.}
\usage{
discharge_parameters(
  start_year = 1951,
  end_year = 2011,
  detrended_directory,
  project_directory
)
}
\arguments{
\item{start_year}{The first year of data to be used.}

\item{end_year}{The last year of data to be used.}

\item{detrended_directory}{The directory where the detrended discharge files are located.}

\item{project_directory}{The directory used to save output files.}
}
\description{
Calculate the annual discharge, the average daily discharge, the 7 day minimum discharge, and the 15 day maximum discharge.
}
