% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_q.R
\name{effects_q}
\alias{effects_q}
\title{Calculate the effects}
\usage{
effects_q(
  project_directory,
  gage_list,
  sum_anova_results,
  mean_anova_results,
  min_anova_results,
  max_anova_results
)
}
\arguments{
\item{project_directory}{The directory used to save output files.}

\item{gage_list}{The gages used in vector form.}

\item{sum_anova_results}{The object returned from the function sum_anova.}

\item{mean_anova_results}{The object returned from the function mean_anova.}

\item{min_anova_results}{The object returned from the function min_anova.}

\item{max_anova_results}{The object returned from the function max_anova.}
}
\description{
Calculate the effects
}
