% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_zero.R
\name{leading_zero}
\alias{leading_zero}
\title{Make sure all gage IDs have leading zeroes.}
\usage{
leading_zero(gage_list)
}
\arguments{
\item{gage_list}{A vector of the gages used.}
}
\description{
Make sure all gage IDs have leading zeroes.
}
