% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowflow_timing.R
\name{lowflow_timing}
\alias{lowflow_timing}
\title{Calculate the most prevalent timing of low flow events between conditions.}
\usage{
lowflow_timing(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\description{
Calculate the most prevalent timing of low flow events between conditions.
}
