% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_kw.R
\name{max_kw}
\alias{max_kw}
\title{ANOVA of the z-scored max discharge values.}
\usage{
max_kw(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\description{
ANOVA of the z-scored max discharge values.
}
