% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_stats.R
\name{max_stats}
\alias{max_stats}
\title{15-Day Maximum Discharge Statistics}
\usage{
max_stats(project_directory, normal = FALSE)
}
\arguments{
\item{project_directory}{The directory used to save output files.}

\item{normal}{TRUE or FALSE will determine if an ANOVA or Kruskal-Wallis test is used}
}
\value{
A list of values
}
\description{
ANOVA or Kruskal-Wallis ranked test of the z-scored max discharge values.
}
