% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_kw.R
\name{mean_kw}
\alias{mean_kw}
\title{ANOVA of the z-scored mean discharge values.}
\usage{
mean_kw(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\description{
ANOVA of the z-scored mean discharge values.
}
