% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_anova.R
\name{min_anova}
\alias{min_anova}
\title{ANOVA of the z-scored min discharge values.}
\usage{
min_anova(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\description{
ANOVA of the z-scored min discharge values.
}
