% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_kw.R
\name{min_kw}
\alias{min_kw}
\title{ANOVA of the z-scored min discharge values.}
\usage{
min_kw(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\description{
ANOVA of the z-scored min discharge values.
}
