% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_stats.R
\name{min_stats}
\alias{min_stats}
\title{7-Day Minimum Discharge Statistics}
\usage{
min_stats(project_directory, normal = FALSE)
}
\arguments{
\item{project_directory}{The directory used to save output files.}

\item{normal}{TRUE or FALSE will determine if an ANOVA or Kruskal-Wallis test is used}
}
\value{
A list of values
}
\description{
ANOVA or Kruskal-Wallis ranked test of the z-scored min discharge values.
}
