% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Calculate the moving average of a set of numbers.}
\usage{
moving_average(dat, avg_length)
}
\arguments{
\item{dat}{A vector of data to create a moving average for.}

\item{avg_length}{The number of values to be averaged.}
}
\description{
Calculate the moving average of a set of numbers.
}
