% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakflow_timing.R
\name{peakflow_timing}
\alias{peakflow_timing}
\title{Calculate the most prevalent timing of peakflow events between conditions.}
\usage{
peakflow_timing(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\description{
Calculate the most prevalent timing of peakflow events between conditions.
}
