% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_trends_histogram.R
\name{plot_slopes}
\alias{plot_slopes}
\title{Histogram of the slopes of each gage's trend in discharge.}
\usage{
plot_slopes(slope_df)
}
\arguments{
\item{slope_df}{Dataframe of the slopes associated with each gage's trend in discharge.}
}
\description{
Histogram of the slopes of each gage's trend in discharge.
}
