% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_all_files.R
\name{pre_process}
\alias{pre_process}
\title{Detrend all of the discharge data for the project.}
\usage{
pre_process(discharge_directory, detrended_directory, project_directory)
}
\arguments{
\item{discharge_directory}{The directory where the raw discharge files are located.}

\item{detrended_directory}{The directory where the detrended discharge files are located.}

\item{project_directory}{The directory used to save output files.}
}
\value{
a list of the gages used for the study
}
\description{
Detrend all of the discharge data for the project.
}
