% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_gages.R
\name{standardize_gages}
\alias{standardize_gages}
\title{Make sure all gage IDs have leading zeroes.}
\usage{
standardize_gages(
  start_year = 1951,
  end_year = 2011,
  project_directory,
  detrended_directory
)
}
\arguments{
\item{start_year}{The first year of data to be used.}

\item{end_year}{The last year of data to be used.}

\item{project_directory}{The directory used to save output files.}

\item{detrended_directory}{The directory where the detrended discharge files are located.}
}
\description{
Make sure all gage IDs have leading zeroes.
}
