% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_anova.R
\name{sum_anova}
\alias{sum_anova}
\title{Total Annual Discharge ANOVA}
\usage{
sum_anova(project_directory)
}
\arguments{
\item{project_directory}{The directory used to save output files.}
}
\value{
A list of values
}
\description{
ANOVA of the z-scored sum discharge values.
}
