% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Z-score}
\usage{
z_score(value, values)
}
\arguments{
\item{value}{the value to z-score.}

\item{values}{the rest of the dataset to z-score against.}
}
\value{
the z-scored value
}
\description{
Z-score a value.
}
